Feature/norminette minishell

+++ srcs/redirections/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: msakata <msakata@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/17 00:00:00 by student           #+#    #+#             */
-/*   Updated: 2025/11/17 13:11:42 by msakata          ###   ########TOKYO.jp  */
+/*   Updated: 2025/11/22 05:31:59 by msakata          ###   ########TOKYO.jp  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,19 +24,10 @@ static char	*get_heredoc_filename(void)
 	return (filename);
 }
 
-int	handle_heredoc(char *delimiter)
+static void	read_heredoc_lines(int fd, char *delimiter)
 {
 	char	*line;
-	char	*filename;
-	int		fd;
 
-	filename = get_heredoc_filename();
-	fd = open(filename, O_WRONLY | O_CREAT | O_TRUNC, 0600);
-	if (fd < 0)
-	{
-		free(filename);
-		return (-1);
-	}
 	while (1)
 	{
 		line = readline("> ");
@@ -48,6 +39,21 @@ int	handle_heredoc(char *delimiter)
 		ft_putendl_fd(line, fd);
 		free(line);
 	}
+}
+
+int	handle_heredoc(char *delimiter)
+{
+	char	*filename;
+	int		fd;
+
+	filename = get_heredoc_filename();
+	fd = open(filename, O_WRONLY | O_CREAT | O_TRUNC, 0600);
+	if (fd < 0)
+	{
+		free(filename);
+		return (-1);
+	}
+	read_heredoc_lines(fd, delimiter);
 	close(fd);
 	fd = open(filename, O_RDONLY);
 	unlink(filename);
